Extract compose widgets
